+++ unix/xft.c
@@ -1729,7 +1729,7 @@ static void
 overstrike( Handle self, int x, int y, Point *ovx, int advance)
 {
 	DEFXX;
-	int lw = apc_gp_get_line_width( self);
+	float lw = apc_gp_get_line_width( self);
 	int d  = - PDrawable(self)-> font. descent;
 	int ay, x1, y1, x2, y2;
 	double c = XX-> xft_font_cos, s = XX-> xft_font_sin;
@@ -1740,8 +1740,8 @@ overstrike( Handle self, int x, int y, Point *ovx, int advance)
 		XX-> flags. brush_fore = 1;
 	}
 
-	if ( lw != 1)
-		apc_gp_set_line_width( self, 1);
+	if ( lw != 1.0)
+		apc_gp_set_line_width( self, 1.0);
 
 	if ( ovx->x < 0 ) ovx->x = 0;
 	if ( ovx->y < 0 ) ovx->y = 0;
@@ -1765,7 +1765,7 @@ overstrike( Handle self, int x, int y, Point *ovx, int advance)
 		XDrawLine( DISP, XX-> gdrawable, XX-> gc, x1, REVERT( y1), x2, REVERT( y2));
 	}
 
-	if ( lw != 1)
+	if ( lw != 1.0)
 		apc_gp_set_line_width( self, lw);
 }
